Refactor project to rename from FastF1 Portfolio to FastF1 Analytics
- Updated project name and references throughout the codebase to reflect the new branding.
- Modified configuration files, documentation, and source links to align with the new project name.
- Adjusted gallery and API documentation to ensure consistency with the updated naming conventions.
- Enhanced import paths and script references to use the new package structure.
