+++ docs/api/plotting.md
@@ -1,9 +1,9 @@
 ﻿# API: plotting
-::: fastf1_portfolio.plotting
+::: fastf1_analytics.plotting
 
 **Examples**
 ```python
-from fastf1_portfolio.plotting import get_team_color, get_compound_color, savefig, apply_style
+from fastf1_analytics.plotting import get_team_color, get_compound_color, savefig, apply_style
 apply_style()
 print(get_team_color("Ferrari"))
 print(get_compound_color("Soft"))
