+++ src/fastf1_analytics/cli.py
