+++ src/fastf1_analytics/session_loader.py
