+++ tests/test_imports.py
@@ -1,7 +1,7 @@
 ﻿def test_package_imports() -> None:
     """Test that the package can be imported and has expected attributes."""
-    import fastf1_portfolio
+    import fastf1_analytics
 
-    assert hasattr(fastf1_portfolio, "load_session")
-    assert hasattr(fastf1_portfolio, "apply_style")
-    assert hasattr(fastf1_portfolio, "__version__")
+    assert hasattr(fastf1_analytics, "load_session")
+    assert hasattr(fastf1_analytics, "apply_style")
+    assert hasattr(fastf1_analytics, "__version__")
